Update cleaningSchools script
In the last step where we are taking the county with the highest population, we are more interested in the student population, not the student body plus faculty, staff, etc. so I changed it from POPULATION to TOT_ENROLL (total enrolled). The results are exactly the same but this is more consistent with the angle we've been taking all along (with the focus on undergrads).

Also changed the line that wrote the csv, I was using the wrong function.